refactor: get rid of duplicated code
